+++ test/strategy.profile.test.js
@@ -2,7 +2,7 @@
 /* jshint expr: true */
 
 var FacebookStrategy = require('../lib/strategy')
-  , graphApiVersion = FacebookStrategy.defaultGraphApiVersion;
+  , graphApiVersion = require('./graphApiVersion')
 
 
 describe('Strategy#userProfile', function() {
@@ -10,7 +10,8 @@ describe('Strategy#userProfile', function() {
   describe('fetched from default endpoint', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
-        clientSecret: 'secret'
+        clientSecret: 'secret',
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -60,7 +61,8 @@ describe('Strategy#userProfile', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
         clientSecret: 'secret',
-        enableProof: true
+        enableProof: true,
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -93,7 +95,8 @@ describe('Strategy#userProfile', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
         clientSecret: 'secret',
-        profileFields: ['id', 'username', 'displayName', 'name', 'gender', 'profileUrl', 'emails', 'photos']
+        profileFields: ['id', 'username', 'displayName', 'name', 'gender', 'profileUrl', 'emails', 'photos'],
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -127,7 +130,8 @@ describe('Strategy#userProfile', function() {
         clientID: 'ABC123',
         clientSecret: 'secret',
         profileFields: ['id', 'username', 'displayName', 'name', 'gender', 'profileUrl', 'emails', 'photos'],
-        enableProof: true
+        enableProof: true,
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -160,7 +164,8 @@ describe('Strategy#userProfile', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
         clientSecret: 'secret',
-        profileFields: ['id', 'username', 'displayName', 'name', 'gender', 'profileUrl', 'emails', 'photos', 'public_key', 'updated_time']
+        profileFields: ['id', 'username', 'displayName', 'name', 'gender', 'profileUrl', 'emails', 'photos', 'public_key', 'updated_time'],
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -197,7 +202,8 @@ describe('Strategy#userProfile', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
         clientSecret: 'secret',
-        profileFields: []
+        profileFields: [],
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -226,144 +232,11 @@ describe('Strategy#userProfile', function() {
     });
   }); // fetched from default endpoint, with profile fields being an empty array
 
-  describe('fetched from older version with fields specified in URL', function() {
-    var strategy = new FacebookStrategy({
-        clientID: 'ABC123',
-        clientSecret: 'secret',
-        profileURL: 'https://graph.facebook.com/v2.2/me?fields=id,username'
-      }, function() {});
-
-    strategy._oauth2.get = function(url, accessToken, callback) {
-      if (url != 'https://graph.facebook.com/v2.2/me?fields=id,username') { return callback(new Error('incorrect url argument')); }
-      if (accessToken != 'token') { return callback(new Error('incorrect token argument')); }
-
-      var body = '{"id":"500308595","name":"Jared Hanson","first_name":"Jared","last_name":"Hanson","link":"http:\\/\\/www.facebook.com\\/jaredhanson","username":"jaredhanson","gender":"male","email":"jaredhanson\\u0040example.com"}';
-      callback(null, body, undefined);
-    };
-
-
-    var profile;
-
-    before(function(done) {
-      strategy.userProfile('token', function(err, p) {
-        if (err) { return done(err); }
-        profile = p;
-        done();
-      });
-    });
-
-    it('should parse profile', function() {
-      expect(profile.provider).to.equal('facebook');
-      expect(profile.id).to.equal('500308595');
-      expect(profile.username).to.equal('jaredhanson');
-    });
-  }); // fetched from older version with fields specified in URL
-
-  describe('fetched from older version with fields specified in URL, with appsecret_proof', function() {
-    var strategy = new FacebookStrategy({
-        clientID: 'ABC123',
-        clientSecret: 'secret',
-        profileURL: 'https://graph.facebook.com/v2.2/me?fields=id,username',
-        enableProof: true
-      }, function() {});
-
-    strategy._oauth2.get = function(url, accessToken, callback) {
-      if (url != 'https://graph.facebook.com/v2.2/me?fields=id,username&appsecret_proof=e941110e3d2bfe82621f0e3e1434730d7305d106c5f68c87165d0b27a4611a4a') { return callback(new Error('incorrect url argument')); }
-      if (accessToken != 'token') { return callback(new Error('incorrect token argument')); }
-
-      var body = '{"id":"500308595","name":"Jared Hanson","first_name":"Jared","last_name":"Hanson","link":"http:\\/\\/www.facebook.com\\/jaredhanson","username":"jaredhanson","gender":"male","email":"jaredhanson\\u0040example.com"}';
-      callback(null, body, undefined);
-    };
-
-
-    var profile;
-
-    before(function(done) {
-      strategy.userProfile('token', function(err, p) {
-        if (err) { return done(err); }
-        profile = p;
-        done();
-      });
-    });
-
-    it('should parse profile', function() {
-      expect(profile.provider).to.equal('facebook');
-      expect(profile.id).to.equal('500308595');
-      expect(profile.username).to.equal('jaredhanson');
-    });
-  }); // fetched from older version with fields specified in URL
-
-  describe('fetched from older version specified in graphApiVersion option', function() {
-    var strategy = new FacebookStrategy({
-        clientID: 'ABC123',
-        clientSecret: 'secret',
-        graphApiVersion: 'v2.2'
-      }, function() {});
-
-    strategy._oauth2.get = function(url, accessToken, callback) {
-      if (url != 'https://graph.facebook.com/v2.2/me') { return callback(new Error('incorrect url argument')); }
-      if (accessToken != 'token') { return callback(new Error('incorrect token argument')); }
-
-      var body = '{"id":"500308595","name":"Jared Hanson","first_name":"Jared","last_name":"Hanson","link":"http:\\/\\/www.facebook.com\\/jaredhanson","username":"jaredhanson","gender":"male","email":"jaredhanson\\u0040example.com"}';
-      callback(null, body, undefined);
-    };
-
-
-    var profile;
-
-    before(function(done) {
-      strategy.userProfile('token', function(err, p) {
-        if (err) { return done(err); }
-        profile = p;
-        done();
-      });
-    });
-
-    it('should parse profile', function() {
-      expect(profile.provider).to.equal('facebook');
-      expect(profile.id).to.equal('500308595');
-      expect(profile.username).to.equal('jaredhanson');
-    });
-  }); // fetched from older version with fields specified in URL
-
-  describe('fetched from older version specified in graphApiVersion option, with appsecret_proof', function() {
-    var strategy = new FacebookStrategy({
-        clientID: 'ABC123',
-        clientSecret: 'secret',
-        graphApiVersion: 'v2.2',
-        enableProof: true
-      }, function() {});
-
-    strategy._oauth2.get = function(url, accessToken, callback) {
-      if (url != 'https://graph.facebook.com/v2.2/me?appsecret_proof=e941110e3d2bfe82621f0e3e1434730d7305d106c5f68c87165d0b27a4611a4a') { return callback(new Error('incorrect url argument')); }
-      if (accessToken != 'token') { return callback(new Error('incorrect token argument')); }
-
-      var body = '{"id":"500308595","name":"Jared Hanson","first_name":"Jared","last_name":"Hanson","link":"http:\\/\\/www.facebook.com\\/jaredhanson","username":"jaredhanson","gender":"male","email":"jaredhanson\\u0040example.com"}';
-      callback(null, body, undefined);
-    };
-
-
-    var profile;
-
-    before(function(done) {
-      strategy.userProfile('token', function(err, p) {
-        if (err) { return done(err); }
-        profile = p;
-        done();
-      });
-    });
-
-    it('should parse profile', function() {
-      expect(profile.provider).to.equal('facebook');
-      expect(profile.id).to.equal('500308595');
-      expect(profile.username).to.equal('jaredhanson');
-    });
-  }); // fetched from older version with fields specified in URL
-
   describe('error caused by invalid token', function() {
     var strategy =  new FacebookStrategy({
         clientID: 'ABC123',
-        clientSecret: 'secret'
+        clientSecret: 'secret',
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -395,7 +268,8 @@ describe('Strategy#userProfile', function() {
   describe('error caused by malformed response', function() {
     var strategy =  new FacebookStrategy({
         clientID: 'ABC123',
-        clientSecret: 'secret'
+        clientSecret: 'secret',
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {
@@ -423,7 +297,8 @@ describe('Strategy#userProfile', function() {
   describe('internal error', function() {
     var strategy = new FacebookStrategy({
         clientID: 'ABC123',
-        clientSecret: 'secret'
+        clientSecret: 'secret',
+        graphApiVersion: graphApiVersion
       }, function() {});
 
     strategy._oauth2.get = function(url, accessToken, callback) {